Add new feature WL: configuring required/optional criteria for vacancy templ sourceforge feature ID: 3602236, link:https://sourceforge.net/tracker/?func=detail&aid=3602236&group_id=66701&atid=515435

+++ database/mysql/oscarinit.sql
@@ -9517,7 +9517,7 @@ CREATE TABLE `criteria` (
   `TEMPLATE_ID` int(11),
   `VACANCY_ID` int(11),
   `MATCH_SCORE_WEIGHT` double NOT NULL,
-  `CAN_BE_ADHOC` tinyint(1) NOT NULL,
+  `CAN_BE_ADHOC` int(1) NOT NULL,
   PRIMARY KEY (`CRITERIA_ID`)
 );
